conversion-container/pyproject.toml
[tool.poetry]
name = "conversion"
version = "1.1.0"
description = "This container runs in cloud run and handles conversion requests and submission publication events"
authors = ["Mark Nazzaro <marknazzaro2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
arxiv-base = {git = "https://github.com/arXiv/arxiv-base.git", rev = "3bdc3161bd10322cd4581db50a7f5ef06663ed85"}
gunicorn = "^20.1.0"
bs4 = "^0.0.1"
filelock = "^3.12.2"
requests = "^2.31.0"
tenacity = "*"
sqlalchemy = "~=2.0.27"
mypy = "*"

[tool.poetry.dev-dependencies]
pytest-mock = "*"

[tool.pytest.ini_options]
markers = [
    "cc_unit_tests: Unit tests for concurrency control functions",
    "processing_unit_tests: Unit tests for processing functions"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
